Crate bp_rialto

Re-exports

pub use time_units::*;

Modules

time_units

Human readable time units defined in terms of number of blocks.

Structs

AccountIdConverter

Convert a 256-bit hash into an AccountId.

BlockLength

BlockWeights

Rialto

Rialto chain.

Constants

ADDITIONAL_MESSAGE_BYTE_DELIVERY_WEIGHT

Increase of delivery transaction weight on Rialto chain with every additional message byte.

AVERAGE_ON_INITIALIZE_RATIO

Represents the average portion of a block’s weight that will be used by an on_initialize() runtime call.

BEST_FINALIZED_RIALTO_HEADER_METHOD

Name of the RialtoFinalityApi::best_finalized runtime method.

DEFAULT_MESSAGE_DELIVERY_TX_WEIGHT

Weight of single regular message delivery transaction on Rialto chain.

EXTRA_STORAGE_PROOF_SIZE

Number of extra bytes (excluding size of storage value itself) of storage proof, built at Rialto chain. This mostly depends on number of entries (and their density) in the storage trie. Some reserve is reserved to account future chain growth.

FROM_RIALTO_LATEST_CONFIRMED_NONCE_METHOD

Name of the FromRialtoInboundLaneApi::latest_onfirmed_nonce runtime method.

FROM_RIALTO_LATEST_RECEIVED_NONCE_METHOD

Name of the FromRialtoInboundLaneApi::latest_received_nonce runtime method.

FROM_RIALTO_UNREWARDED_RELAYERS_STATE

Name of the FromRialtoInboundLaneApi::unrewarded_relayers_state runtime method.

MAXIMAL_ENCODED_ACCOUNT_ID_SIZE

Maximal size (in bytes) of encoded (using Encode::encode()) account id.

MAXIMUM_BLOCK_WEIGHT

Maximal weight of single Rialto block.

MAX_SINGLE_MESSAGE_DELIVERY_CONFIRMATION_TX_WEIGHT

Maximal weight of single message delivery confirmation transaction on Rialto chain.

MAX_UNCONFIRMED_MESSAGES_AT_INBOUND_LANE

Maximal number of unconfirmed messages at inbound lane.

MAX_UNREWARDED_RELAYER_ENTRIES_AT_INBOUND_LANE

Maximal number of unrewarded relayer entries at inbound lane.

NORMAL_DISPATCH_RATIO

Represents the portion of a block that will be used by Normal extrinsics.

PAY_INBOUND_DISPATCH_FEE_WEIGHT

Weight of pay-dispatch-fee operation for inbound messages at Rialto chain.

SESSION_LENGTH

The target length of a session (how often authorities change) on Rialto measured in of number of blocks.

TO_RIALTO_ESTIMATE_MESSAGE_FEE_METHOD

Name of the ToRialtoOutboundLaneApi::estimate_message_delivery_and_dispatch_fee runtime method.

TO_RIALTO_LATEST_GENERATED_NONCE_METHOD

Name of the ToRialtoOutboundLaneApi::latest_generated_nonce runtime method.

TO_RIALTO_LATEST_RECEIVED_NONCE_METHOD

Name of the ToRialtoOutboundLaneApi::latest_received_nonce runtime method.

TO_RIALTO_MESSAGE_DETAILS_METHOD

Name of the ToRialtoOutboundLaneApi::message_details runtime method.

TX_EXTRA_BYTES

Number of bytes, included in the signed Rialto transaction apart from the encoded call itself.

Traits

FromRialtoInboundLaneApi

Inbound message lane API for messages sent by Rialto chain.

RialtoFinalityApi

API for querying information about the finalized Rialto headers.

ToRialtoOutboundLaneApi

Outbound message lane API for messages that are sent to Rialto chain.

Functions

derive_account_from_millau_id

max_extrinsic_size

Get the maximum length in bytes that a Normal extrinsic on the Millau chain requires.

max_extrinsic_weight

Get the maximum weight (compute time) that a Normal extrinsic on the Millau chain can use.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountSigner

Public key of the chain account that may be used to verify signatures.

Balance

Balance of an account.

BlockNumber

Block number type used in Rialto.

Hash

Hash type used in Rialto.

Hasher

The type of an object that can produce hashes on Rialto.

Header

The header type used by Rialto.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.